fix of erroneous definition of legexpansion()
within legexpansion() the distinction between theta_i, theta_ex and theta_s was not done correctly. the legexpansion-function is different for volume.py and surface.py which was not taken care of previously.

since calculating the fn-coefficients in most general way (i.e. as a function of theta_i, theta_ex, phi_i, phi_ex) is very time-consuming and in most cases not necessary, an additional input-parameter called 'geometry' was added. It is defined as a 4-character string that defines the geometry of the measurement-setup. It's use is explained in the comment-section of the legexpansion() function.