fix of erroreous definition of legexpansion()
the angles used as arguments in thetap and thetabrdf are the zenith-angles while the angles used as arguments for the phase-functions within the paper are the corresponding angles in a spherical coordinate system! this has not been taken into account for the legexpansion() in volume.py but it has already been accounted within surface.py. this inconsistency is now fixed.