move of omega-tau assert from Volume to RT1
in order to be able to call volume-class elements without specifying omega and tau (needed to be able to define linearcombinations without omega-tau ambiguities), the asserts have been moved to the RT1-class. This is reasonable, since omega and tau are only needed for the calculations within the RT1-class.