fix of ill-defined mono-geometry specification
specifying geometry='mono' did not affect the results of Isurf and Ivol since they have been calculated with respect to the given numerical values for mu_ex and phi_ex which possibly results in an erroreous mixing of different geometries. furthermore the assignments for phi_0 and phi_ex have also been fixed.